requirements.txt
# Core LangChain
langchain>=0.3.0

# Ngrok tunnel (for local development)
ngrok>=0.0.0
langchain-core>=0.3.0

# Free LLM providers (choose one or both)
langchain-groq>=0.2.0          # Groq free API  — recommended
langchain-ollama>=0.2.0        # Ollama local   — optional fallback

# Output parsing & env
pydantic>=2.0.0
python-dotenv>=1.0.0

# Google Calendar
google-api-python-client>=2.120.0
google-auth-httplib2>=0.2.0
google-auth-oauthlib>=1.2.0

# Async dependencies
aiosqlite>=0.19.0
httpx>=0.25.0

# Flask
Flask>=3.1.3

# Scheduling optimisation (Flexible Job Shop / CP-SAT)
ortools>=9.8
